FirebaseAuthContext: Remove unneeded method
Remove the user creation function as it is no longer required. This allowed user creation from from the frontend but no control over custom claims.
We have instead implemented a user creation function that consists of an api call where the creation and setting of custom claims takes place in the backend.