controller: Set upline uid when creating new user
Get all user ids from Firestore, then check if the referrer id that was passed in as a param is amongst them.
If it is, we know it is valid so can set it as the new users upline uid. Can also get the uplines role from Firestore.
If it is not, the upline uid can be set to the admin uid. That will be the case of a user that has not been referred or an incorrect referrer uid.