controller: Add init function
Add a function that will be used to initialize payments for a new user. It will (for example) set payments to zero so that a record exists and can be added to.
Without this, we would have to check each field in the 'create' function as they may be null which would cause an error.