controller: Write init function
Initialize all stats so they are zero. This way the 'create' function can add to them, knowing they are a number without checking first.
This will reduce the size of the 'create' function that will be very large in any case. It is also more computationally efficient as checks no longer have to be run every time a request is made to 'create'.