paymentsRoute: Remove uid param
No need to pass the current users id in as a param. In fact, using this value is unsafe as a stolen uid could be passed in by the caller. They would however already have to be a member.
Will instead use value from locals which comes from the auth token and does not depend on user input.