controller: Add 'subscribed' custom claim
When creating a new user, add a new custom claim called 'subscribed' that is initialized to false.
This tracks if the user has paid as a subscriber before (i.e. with recurring payments). It will be set to true once payment is made so that subsequent payments do not get added to MRR.
Also move the declaration of variables from the request body to the top.