controller: Improve Firestore call in 'invoices'
Redesign the call to Firestore so that only a single request is made rather than looping through the invoice ids and making a separate call each time.
This reduced the total response time from approx 30 secs to 9 secs. The saving would increase linearly as the number of invoices grew in number from the current few dozen.