ActivityListView: Make rawReceipts a variable
Instead of having rawReceipts as state, make it the returned value from the getData function.
This allows the receipts data to be displayed in the modal as desired. Implementing as previously where getData sets the state caused a bug whereby the data did not populate until the second opening of the modal.