CustomerList: Reposition 'role' column
Move the role column from position 3 to position 2.
One reason is that it is better suited beside the user name and email.
The other is that now that column widths can be resized manually but are the same at start, it makes sense to have an optimally wider column (name) next to an optimally narrower column (role). A single drag by the user will then optimize layout.