ReceiptsModal: Adjust element layering
Increase the z-index of the modal overlay to 1000 as the column dividers had an index of 999 for some reason.
Table header also had a z-index of 99 so need to be on front of that too.
The MatX sidebar and topbar also have a high z-index so good to be on front of those too.