controller: Remove stats initialization from /payments
Had initialized stats when a payment was first made. However, these conditionals get checked every time the user makes a payment - would be better to run once only on account creation.
This also reduces the size and runtime of the overly-large 'create' function here. It also means that users that need stats (such as admin) do not need to make an unnecessary payment to init them.