controller: Add stats initialization to /users
Initialize stats in the `create` function for /users rather than /payments.
This reduces the amount of times it needs to be run (now once only). It also eliminates the need for a costly await call to Firestore as was necessary in /payments (since we know stats is empty, no need to check).