controller: Remove name from stats
No need to initialize stats with a name. The name is set by the user later so it would need to be updated here and in user too. Better to have a single source of truth.
Uid and email are fine though as they do not change and are known from the start.