controller: Change the order of stats setting
Set the admin stats first, then the stats for everyone.
This is because if the admin is done first, only they await the response. The other users can then set their their stats without awaiting.
This is better because the other users will be set much more frequently than the admin.