Prevent scrollbar rail from overlapping list items
Reduce the padding of the receiptsModal by 18mm and increase the side margin of the ReceiptsListView list items (receipts) by the same amount.
This has the effect of moving the Scrollbar rail closer to the side of the modal. It hugs the modal padding, and therefore doesn't overlap the list items anymore.