controller: Save avatarUrl to Firestore in 'edit'
Save the avatarUrl to Firestore as well as auth. In some locations, the user data is only gathered from Firestore.
Named photoURL in Firebase auth as the field names are fixed. Good to have a different name because although they point to the same resource, they are stored in different locations.