InvoicesList: Add user name and avatarUrl to invoice data
Needed to add the avatarUrl so that photo is displayed next to each list item. This could have been added to the invoice data in Firestore on creation to save recomputing every time. But this would not backfill past invoices.
It is for this reason that name was added back in here too. Now name is displayed instead of email for items that predate the user adding their name.