MatxCustomizer: Update how sidebar closes
The previous implementation had local state determining whether the sidebar should be open or not. Closing it resulted in going back to <SecondarySidebarContent/>. Now we close by turning off the sidebar in settings.
Also remove <Tooltip/> and nested <Label/>. These appeared on the intermediate <SecondarySidebarContent/> component that is being removed.