Delete all Drag&Drop
This may be kind of useful for when the drag-and-drop functionality is being introduced to the Put Gang genealogy tree. But there is so much irrelevant code and needless complexity, it would be better to just import react-beautiful-dnd (which it utilizes) and go from there.
Could also use react-dnd, which offers more fine-grained control. Hopefully neither is needed though, and a sortable tree can provide the needed functionality out of the box instead.