instagramAvatar: Remove functions from prior Instagram api attempt
Remove the getAvatar function and it's four helpers from this file. These were all the functions associated with the previous attempt to get the Instagram api working on a prior branch.
The problem with the previous solution was that it did not work in the remote environment. The new solution uses programmatic login to Instagram and CSRF tokens to gain access.