Charts: Add user data from rawData
Add the user data along with the revenue data. Both graphs in this section use the same api so they can use data formatted in the same way.
This is preferable to the alternative, which is writing a separate function that would do much of the same operations on the data.