helpers: Add recurringPayment function
Write the recurringPayment function. It's basically a copy of the 'create' function from /payment for now.
Differences are that it is not an Express function (no req / res). Also no try / catch and no error handling.  Required variables are passed in as params rather than via the req and res objects.