navigations: Disable linter warning
The linter complained that authRoles were defined but never used. Add a comment to silence this warning.
It was not clear how to implement authRoles. However, the MatX team have given access to their GitHub repo with the latest version where this should be more clear. Will implement to conditionally render charts when there is time.