controller: Check paying customer with 'free' in 'create'
Instead of checking 'membLvl' before iterating the customer in charts, check the 'free' variable that sometimes gets passed in the request body.
Nothing gets passed when creating a user the usual way, so it evaluates to false and the function runs. But with a free tier user, it will be passed.