constants: Correct admin uid
Needed to generate a new admin uid. This is because the admin needs to be created in Firebase after the email has been set here. Only then can a uid be generated and set.

Setting both at the same time as before resulted in the admin getting erroneously established as a standard user.